requirements.txt
http://download.pytorch.org/whl/cpu/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
flask==1.0.2
pillow==4.3.0
fastai==1.0.57